Move (unmodified) download functions to utils/sources.py
